refactor: Change return type of 'todo' method in Dispatcher to double
Modify the 'todo' method in the 'Dispatcher' class to change its return type from void to double. This allows the method to return a value that is calculated within a chain of if statements.

Also, introduce a new variable to store the value returned by the chain of if statements.

#Dispatcher #refactor #returntype #ifstatements